meato/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-countries==7.3.2
djangorestframework==3.13.1
Faker==14.2.0
model-bakery==1.7.0
Pillow==9.2.0
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
